requirements.txt
Flask
pytesseract
pdf2image
Pillow
requests
opencv-python
google-cloud-translate
